backend/utils/calendar_helper.py
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
from config.settings import settings
from typing import Dict, Optional
import json

async def create_calendar_event(
    title: str,
    description: str,
    start_time: str,
    end_time: str,
    location: Optional[str] = None
) -> Dict:
    """Create Google Calendar event"""
    
    # Mock implementation - replace with actual Google Calendar API
    if settings.GOOGLE_CALENDAR_CREDENTIALS:
        try:
            # Load credentials
            # creds = Credentials.from_authorized_user_file(settings.GOOGLE_CALENDAR_CREDENTIALS)
            # service = build('calendar', 'v3', credentials=creds)
            
            # event = {
            #     'summary': title,
            #     'location': location,
            #     'description': description,
            #     'start': {
            #         'dateTime': start_time,
            #         'timeZone': 'Asia/Kolkata',
            #     },
            #     'end': {
            #         'dateTime': end_time,
            #         'timeZone': 'Asia/Kolkata',
            #     },
            #     'reminders': {
            #         'useDefault': False,
            #         'overrides': [
            #             {'method': 'popup', 'minutes': 30},
            #         ],
            #     },
            # }
            
            # created_event = service.events().insert(calendarId='primary', body=event).execute()
            
            return {
                "id": "mock_event_id",
                "status": "created",
                "link": "https://calendar.google.com/event",
                "provider": "google_calendar"
            }
        except Exception as e:
            print(f"Calendar event creation error: {e}")
            return {
                "status": "error",
                "error": str(e)
            }
    else:
        # Mock response
        print(f"Mock Calendar Event: {title} from {start_time} to {end_time}")
        return {
            "id": "mock_event_id",
            "status": "mock_created",
            "title": title,
            "start_time": start_time,
            "end_time": end_time,
            "provider": "mock"
        }
