backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
aiofiles==23.2.1
pillow==10.2.0
reportlab==4.0.9
openai==1.10.0
google-generativeai==0.3.2
requests==2.31.0
httpx==0.26.0
